filter/mancompress: fix functionality, use xz

+++ filter/mancompress.sh
@@ -4,17 +4,23 @@
 ((ABMANCOMPRESS)) || return
 set_opt globstar
 __mancomp_todo=()
-for i in "$PKGDIR"/usr/share/man/**/*.[0-9nl]; do
-  if [[ -f $i ]]; then
-    __mancomp_todo+=("$i")
-  elif [[ -L $i ]]; then
-    __mancomp_lnk=$(readlink  -- "$i")
-    rm "$i"
-    ln -sf "$__mancomp_lnk".gz "$i"
-  else
-    abwarn "abmancomp WTF for ${i#"$PKGDIR"}"
-  fi
-done
-gzip -9 -- "${__to_comp[@]}"
+
+if [ -d "$PKGDIR"/usr/share/man ]; then
+	for i in "$PKGDIR"/usr/share/man/**/*.*; do
+		if [[ -L $i ]]; then
+			__mancomp_lnk=$(namei "$i" | tail -1 | awk '{print $NF}')
+			rm "$i"
+			ln -sf "$__mancomp_lnk".xz "$i"
+		elif [[ -f $i ]]; then
+			__mancomp_todo+=("$i")
+		else
+			abwarn "abmancomp WTF for ${i#"$PKGDIR"}"
+		fi
+	done
+
+	abinfo "Compressing man page ${__mancomp_todo[@]} ..."
+	xz -9e -- "${__mancomp_todo[@]}"
+fi
+
 unset __mancomp_todo __mancomp_lnk
 rec_opt globstar
